netdog: Move bond-related config structures to their own module
The bond-related config structures have always lived inside the
device-related bond module in `net_config`.  They are used in
`net_config` but also re-exported and used in the `wicked` module.  This
commit moves these structures to their own module.  Doing so makes the
import paths of these structs cleaner and makes the overall import
structure more sane.
