+++ sources/api/netdog/src/wicked/mod.rs
@@ -8,8 +8,9 @@ mod dhcp;
 mod static_address;
 mod vlan;
 
+use crate::bonding::BondMonitoringConfigV1;
 use crate::interface_id::{InterfaceId, InterfaceName, MacAddress};
-use crate::net_config::devices::bonding::{BondMonitoringConfigV1, NetBondV1};
+use crate::net_config::devices::bonding::NetBondV1;
 use crate::net_config::devices::interface::NetInterfaceV2;
 use crate::net_config::devices::vlan::NetVlanV1;
 use crate::net_config::devices::NetworkDeviceV1;
